Closes #123 - Made ArrayBytes a private inner class of Bytes.  Moved any conversion code between Bytes and external data formats to ByteUtil.
